requirements.txt
flask~=1.1.2
ics~=0.7.2
python-dateutil~=2.8.2
pdfplumber~=0.9.0
opencv-python~=4.7.0.72
numpy~=1.21.5
matplotlib~=3.5.2
pytesseract~=0.3.10